src/main/java/util/common/StringUtils.java
package util.common;

import java.util.Random;

/**
 * Created by liangjj on 2017/9/13.
 */
class StringUtils {
    private StringUtils() {
    }

    /**
     * 判断字符串是否为空
     *
     * @param str
     * @return
     */
    static boolean isEmpty(String str) {
        return str == null || str.length() <= 0;
    }

    /**
     * 生成随机字符串
     *
     * @param length 表示生成字符串的长度
     * @return
     */
    static String generateRandomString(int length) {
        String base = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
        Random random = new Random();
        StringBuilder stringBuilder = new StringBuilder();
        for (int i = 0; i < length; i++) {
            int number = random.nextInt(base.length());
            // 返回指定索引处的 char 值
            stringBuilder.append(base.charAt(number));
        }
        return stringBuilder.toString();
    }
}
